consolidate code with seed questions
